pyproject.toml
[tool.poetry]
name = "n8creator_gendiff"
version = "0.1.1"
description = ""
authors = ["n8creator <n8creator@pm.me>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
